Refactor integrations
