+++ mln/models/dynamic/dynamic.py
@@ -12,7 +12,7 @@ from django.core.validators import MaxValueValidator
 from django.db import models
 from django.utils.timezone import now
 
-from ..static import Answer, Color, EnumField, ItemInfo, ItemType, MessageBody, Stack, Question
+from ..static import Answer, Color, EnumField, ItemInfo, ItemType, MessageBody, Stack, Question, MessageTemplate
 from ...services.inventory import assert_has_item
 
 DAY = datetime.timedelta(days=1)
@@ -201,3 +201,12 @@ def get_or_none(cls, *args, **kwargs):
 		return cls.objects.get(*args, **kwargs)
 	except ObjectDoesNotExist:
 		return None
+
+class IntegrationMessage(models.Model):
+	template = models.ForeignKey(MessageTemplate, related_name="+", on_delete=models.CASCADE)
+	networker = models.ForeignKey(User, related_name="+", on_delete=models.CASCADE, limit_choices_to={"profile__is_networker": True})
+	award = models.IntegerField()
+	client = models.ForeignKey(OAuthClient, related_name="+", on_delete=models.CASCADE)
+
+	def __str__(self):
+		return f"{self.client.client_name}: Message #{self.award}"
