+++ mln/models/static/static.py
@@ -375,19 +375,3 @@ class NetworkerMessageAttachmentLegacy(Stack):
 
 	class Meta:
 		constraints = (models.UniqueConstraint(fields=("trigger", "item"), name="networker_message_attachment_unique_trigger_item"),)
-
-class CoastGuardMessage(models.Model):
-	template = models.ForeignKey(MessageTemplate, related_name="+", on_delete=models.CASCADE)
-	networker = models.ForeignKey(User, related_name="+", on_delete=models.CASCADE, limit_choices_to={"profile__is_networker": True})
-	rank = models.IntegerField()
-
-	def __str__(self):
-		return f"Rank {self.rank} message"
-
-class RobotChroniclesMessage(models.Model):
-	template = models.ForeignKey(MessageTemplate, related_name="+", on_delete=models.CASCADE)
-	networker = models.ForeignKey(User, related_name="+", on_delete=models.CASCADE, limit_choices_to={"profile__is_networker": True})
-	award = models.IntegerField()
-
-	def __str__(self):
-		return f"Rank {self.rank} message"
