+++ mlnserver/urls.py
@@ -13,7 +13,7 @@ from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.views import PasswordChangeView
 from django.views.generic.edit import CreateView
 
-from mln.services import coast_guard, robot_chronicles
+from mln.services import integrations
 from .oauth import OAuthLoginView, get_token
 
 def flashvars_handler(request):
@@ -38,8 +38,7 @@ urlpatterns = [
 	), name="sign_up"),
 	path("accounts/password_change", PasswordChangeView.as_view()),
 	path("admin/", admin.site.urls),
-	path("api/coast-guard/rank", coast_guard.submit_rank),
-	path("api/robot-chronicles/award", robot_chronicles.get_award),
+	path("api/award", integrations.grant_award)
 ]
 
 if settings.DEBUG:
